fix: or separator and stakeholders colors
