+++ frontend/components/icons/Sprite.tsx
@@ -1732,21 +1732,6 @@ const Sprite = () => (
         fill="#F44336"
       />
     </symbol>
-    <symbol
-      id="or-separator"
-      width="436"
-      height="24"
-      viewBox="0 0 436 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <rect y="11" width="436" height="1" fill="#CBD2D9" />
-      <rect x="194" width="48" height="24" fill="white" />
-      <path
-        d="M210.741 17.168C209.77 17.168 208.921 16.9533 208.193 16.524C207.465 16.0947 206.895 15.502 206.485 14.746C206.074 13.9807 205.869 13.0987 205.869 12.1C205.869 11.1013 206.074 10.224 206.485 9.468C206.895 8.70267 207.465 8.10533 208.193 7.676C208.921 7.24667 209.77 7.032 210.741 7.032C211.711 7.032 212.561 7.24667 213.289 7.676C214.017 8.10533 214.581 8.70267 214.983 9.468C215.393 10.224 215.599 11.1013 215.599 12.1C215.599 13.0987 215.393 13.9807 214.983 14.746C214.581 15.502 214.017 16.0947 213.289 16.524C212.561 16.9533 211.711 17.168 210.741 17.168ZM210.741 15.838C211.413 15.838 211.996 15.6887 212.491 15.39C212.995 15.0913 213.387 14.662 213.667 14.102C213.947 13.542 214.087 12.8747 214.087 12.1C214.087 11.3253 213.947 10.658 213.667 10.098C213.387 9.538 212.995 9.10867 212.491 8.81C211.996 8.51133 211.413 8.362 210.741 8.362C210.069 8.362 209.481 8.51133 208.977 8.81C208.473 9.10867 208.081 9.538 207.801 10.098C207.53 10.658 207.395 11.3253 207.395 12.1C207.395 12.8747 207.53 13.542 207.801 14.102C208.081 14.662 208.473 15.0913 208.977 15.39C209.481 15.6887 210.069 15.838 210.741 15.838ZM217.252 17V7.2H220.626C221.4 7.2 222.035 7.33067 222.53 7.592C223.034 7.85333 223.407 8.208 223.65 8.656C223.892 9.09467 224.014 9.58467 224.014 10.126C224.014 10.7327 223.85 11.2833 223.524 11.778C223.206 12.2727 222.707 12.6273 222.026 12.842L224.112 17H222.39L220.486 13.052H218.736V17H217.252ZM218.736 11.932H220.542C221.214 11.932 221.704 11.7687 222.012 11.442C222.329 11.1153 222.488 10.6907 222.488 10.168C222.488 9.64533 222.334 9.23 222.026 8.922C221.718 8.60467 221.218 8.446 220.528 8.446H218.736V11.932Z"
-        fill="#768291"
-      />
-    </symbol>
 
     {/* Dashboard Blobs */}
     <symbol id="blob-purple" viewBox="0 0 100 100" fill="none" xmlns="http://www.w3.org/2000/svg">
